+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Update internal type for compute plan list (#254)
+
 ## [0.6.0] - 2021-12-01
 
 ### Added